pizza/requirements.txt
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.39.1
django-crispy-forms==1.7.2
djangorestframework==3.10.1
entrypoints==0.3
flake8==3.7.8
idna==2.8
mccabe==0.6.1
oauthlib==3.1.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyrsistent==0.15.4
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
